# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "MovePreferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "move",
            "title": "Kind",
            "type": "string"
          },
          "uci": {
            "title": "Uci",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "uci"
        ],
        "title": "MovePreferenceResponse",
        "type": "object"
      },
      "MutationMovePreferenceRequest": {
        "properties": {
          "kind": {
            "const": "move",
            "title": "Kind",
            "type": "string"
          },
          "uci": {
            "title": "Uci",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "uci"
        ],
        "title": "MutationMovePreferenceRequest",
        "type": "object"
      },
      "MutationNoPreferenceRequest": {
        "properties": {
          "kind": {
            "const": "no_preference",
            "title": "Kind",
            "type": "string"
          },
          "uci": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uci"
          }
        },
        "required": [
          "kind"
        ],
        "title": "MutationNoPreferenceRequest",
        "type": "object"
      },
      "NoPreferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "no_preference",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "NoPreferenceResponse",
        "type": "object"
      },
...
      "PreferredMovesErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "invalid_fen",
              "invalid_from",
              "invalid_until",
              "invalid_window",
              "invalid_effective_from",
              "invalid_effective_until",
              "invalid_preference",
              "invalid_uci",
              "illegal_move",
              "preferred_moves_unavailable",
              "unexpected_failure"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "PreferredMovesErrorResponse",
        "type": "object"
      },
      "PreferredMovesMutationRequest": {
        "properties": {
          "effective_from": {
            "title": "Effective From",
            "type": "string"
          },
          "effective_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Until"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "preference": {
            "discriminator": {
              "mapping": {
                "move": "#/components/schemas/MutationMovePreferenceRequest",
                "no_preference": "#/components/schemas/MutationNoPreferenceRequest"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/MutationMovePreferenceRequest"
              },
              {
                "$ref": "#/components/schemas/MutationNoPreferenceRequest"
              }
            ],
            "title": "Preference"
          }
        },
        "required": [
          "fen",
          "effective_from",
          "preference"
        ],
        "title": "PreferredMovesMutationRequest",
        "type": "object"
      },
      "PreferredMovesMutationResponse": {
        "additionalProperties": false,
        "properties": {
          "effective_from": {
            "title": "Effective From",
            "type": "string"
          },
          "effective_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Until"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/PreferredMovesPeriodResponse"
            },
            "title": "Periods",
            "type": "array"
          },
          "preference": {
            "discriminator": {
              "mapping": {
                "move": "#/components/schemas/MovePreferenceResponse",
                "no_preference": "#/components/schemas/NoPreferenceResponse"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/MovePreferenceResponse"
              },
              {
                "$ref": "#/components/schemas/NoPreferenceResponse"
              }
            ],
            "title": "Preference"
          }
        },
        "required": [
          "fen",
          "effective_from",
          "effective_until",
          "preference",
          "periods"
        ],
        "title": "PreferredMovesMutationResponse",
        "type": "object"
      },
      "PreferredMovesPeriodResponse": {
        "additionalProperties": false,
        "properties": {
          "effective_from": {
            "title": "Effective From",
            "type": "string"
          },
          "effective_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Until"
          },
          "preference": {
            "discriminator": {
              "mapping": {
                "move": "#/components/schemas/MovePreferenceResponse",
                "no_preference": "#/components/schemas/NoPreferenceResponse"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/MovePreferenceResponse"
              },
              {
                "$ref": "#/components/schemas/NoPreferenceResponse"
              }
            ],
            "title": "Preference"
          }
        },
        "required": [
          "effective_from",
          "effective_until",
          "preference"
        ],
        "title": "PreferredMovesPeriodResponse",
        "type": "object"
      },
      "PreferredMovesRemovalRequest": {
        "properties": {
          "effective_from": {
            "title": "Effective From",
            "type": "string"
          },
          "effective_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Until"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          }
        },
        "required": [
          "fen",
          "effective_from"
        ],
        "title": "PreferredMovesRemovalRequest",
        "type": "object"
      },
      "PreferredMovesRemovalResponse": {
        "additionalProperties": false,
        "properties": {
          "effective_from": {
            "title": "Effective From",
            "type": "string"
          },
          "effective_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Until"
          },
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/PreferredMovesPeriodResponse"
            },
            "title": "Periods",
            "type": "array"
          }
        },
        "required": [
          "fen",
          "effective_from",
          "effective_until",
          "periods"
        ],
        "title": "PreferredMovesRemovalResponse",
        "type": "object"
      },
      "PreferredMovesResponse": {
        "additionalProperties": false,
        "properties": {
          "fen": {
            "title": "Fen",
            "type": "string"
          },
          "from": {
            "title": "From",
            "type": "string"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/PreferredMovesSegmentResponse"
            },
            "title": "Segments",
            "type": "array"
          },
          "until": {
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "fen",
          "from",
          "until",
          "segments"
        ],
        "title": "PreferredMovesResponse",
        "type": "object"
      },
      "PreferredMovesSegmentResponse": {
        "additionalProperties": false,
        "properties": {
          "from": {
            "title": "From",
            "type": "string"
          },
          "preference": {
            "discriminator": {
              "mapping": {
                "move": "#/components/schemas/MovePreferenceResponse",
                "no_preference": "#/components/schemas/NoPreferenceResponse",
                "unconfigured": "#/components/schemas/UnconfiguredPreferenceResponse"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/MovePreferenceResponse"
              },
              {
                "$ref": "#/components/schemas/NoPreferenceResponse"
              },
              {
                "$ref": "#/components/schemas/UnconfiguredPreferenceResponse"
              }
            ],
            "title": "Preference"
          },
          "until": {
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "from",
          "until",
          "preference"
        ],
        "title": "PreferredMovesSegmentResponse",
        "type": "object"
      },
      "UnconfiguredPreferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "unconfigured",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "UnconfiguredPreferenceResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Chess Move Trainer",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/preferred-moves": {
      "delete": {
        "operationId": "deletePreferredMoves",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferredMovesRemovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesRemovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Preferred Moves",
        "tags": [
          "preferred-moves"
        ]
      },
      "get": {
        "operationId": "getPreferredMoves",
        "parameters": [
          {
            "in": "query",
            "name": "fen",
            "required": true,
            "schema": {
              "title": "Fen",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": true,
            "schema": {
              "title": "Until",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Preferred Moves",
        "tags": [
          "preferred-moves"
        ]
      },
      "put": {
        "operationId": "putPreferredMoves",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferredMovesMutationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferredMovesErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Put Preferred Moves",
        "tags": [
          "preferred-moves"
        ]
      }
    }
  }
}